Add basic support for mongo testing
This is not intended to be for general consumption at this stage; it
requires that the tester has a mongod instance running and because of
the time taken to load the db, $GOROOT/src/cmd/go/test.go must be
modified to allow for tests longer than 10 minutes. A future CL may
allow db probing and the use of an existing cayley test dataset.