simplify the signature of NextLogOut
by getting the value from it.Result() rather than passing it directly,
this reduces the number of things which need to be passed and may help
identify cases where the value was passed to NextLogOut but was not
set in a way to make it.Result() work as it should. in those cases,
the logging would include a result different to what you would get from
it.Result() but lead you to believe that it.Result() would return the
logged value.