graph: rewrite pk; remove horizon and ts from delta; rewrite memstore
Primary key was broken in many ways, for example it carried mutex,
but all updates happened with Next that copied value (with mutex).
Most backends are able to provide their own pk for log entries, so it
better to use it instead of our own implementation.

Delta was changed in incompatible way: ID and Timestamp was removed.
As described above, ID shold be set by QS after TX, while TS should
not be sent from client. Both values are now set in backend code.

Memstore was rewritten to reflect changes. It was also updated to
use Primitives instead of nodes/quads.