Rewrite CockroachDB SQL construction to be much faster.
Because CockroachDB is a distributed database, round trips to it are
much slower than comparable round trips in, say, Postgres, so different
tradeoffs are required for optimal insertions. In this case, a single
INSERT statement with many values is much faster for CockroachDB to
handle than a sequence of single value INSERT statements.

In one test case involving loading a large number of nodes and quads,
this change produces a 4-5x speedup.

See https://github.com/cockroachdb/cockroach/issues/17108 for a more
detailed discussion of this matter.

Fixes #691.