refactor: return errors from ValueOf, NameOf, QuadDirection
Throughout the code there are calls to ValueOf() and NameOf() on the QuadStore.
The QuadStore currently returns a single value for these functions without an
error, but internally the QuadStore might run into some errors when processing
the value. This hack is clearly identified in the code: rather than return the
error, it's just logged.

If the QuadStore has an error while processing ValueOf or NameOf, the returned
value will be nil which might lead to many types of undefined behavior - such as
queries returning "nil" for one field of a Quad.

Fixes #985

Signed-off-by: Christian Stewart <christian@paral.in>